Partial modularization

+++ src/Base/Terrain.cpp
@@ -3,12 +3,13 @@
 #include <iostream>
 
 #include "Terrain.h"
-#include "Hierarchy.h"
+import Hierarchy;
+
 #include "Camera.h"
 #include "SLK.h"
 #include "HiveWE.h"
-#include "BinaryReader.h"
-#include "BinaryWriter.h"
+import BinaryReader;
+import BinaryWriter;
 #include "Physics.h"
 
 using namespace std::literals::string_literals;
@@ -175,7 +176,7 @@ void Terrain::create() {
 	// Ground textures
 	for (const auto& tile_id : tileset_ids) {
 		ground_textures.push_back(resource_manager.load<GroundTexture>(terrain_slk.data("dir", tile_id) + "/" + terrain_slk.data("file", tile_id)));
-		ground_texture_to_id.emplace(tile_id, ground_textures.size() - 1);
+		ground_texture_to_id.emplace(tile_id, static_cast<int>(ground_textures.size() - 1));
 	}
 	blight_texture = static_cast<int>(ground_textures.size());
 	ground_texture_to_id.emplace("blight", blight_texture);
@@ -447,7 +448,7 @@ void Terrain::change_tileset(const std::vector<std::string>& new_tileset_ids, st
 
 	for (const auto& tile_id : tileset_ids) {
 		ground_textures.push_back(resource_manager.load<GroundTexture>(terrain_slk.data("dir", tile_id) + "/" + terrain_slk.data("file", tile_id) + (hierarchy.hd ? "_diffuse.dds" : ".dds")));
-		ground_texture_to_id.emplace(tile_id, ground_textures.size() - 1);
+		ground_texture_to_id.emplace(tile_id, static_cast<int>(ground_textures.size() - 1));
 	}
 	blight_texture = static_cast<int>(ground_textures.size());
 	ground_texture_to_id.emplace("blight", blight_texture);
